requirements.txt
Flask==3.0.2
IMDbPY==2022.7.9
Werkzeug==3.0.1
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
python-dotenv==1.0.1
